Ensure that the channel name doesn't contain dots when passed to Faye
There was a bug where repos with multiples dots were throwing an exception
