+++ ember-app/app/initializers/deferReadiness.js
@@ -1,0 +1,7 @@
+export function initialize(container, application){
+  application.deferReadiness();
+}
+export default {
+  name: 'deferReadiness',
+  initialize: initialize
+}
